[REF] web_studio: rework the way studio: 1 is set in context
Before this commit, in studio, we created a sub env with a custom
user service, where the user context always contained a key studio,
which altered the behavior of some python methods for studio needs.

This is no longer possible as user is no longer a service. It can
thus no longer be specialized/customized for a sub component-tree.

This commit reworks the logic to make this work again. View editors
are spawned with the key in their context, such that all rpcs they
do correctly have the context key. Studio routes also force the
key in the context, such that all python methods called from them
have the context key as well.

Part of task~3621046

Part-of: odoo/enterprise#52372