[IMP] account_reports: better multi-company tax closing
When performing the tax closing in a multi-company environment (either because of a tax unit or branches), the tax report pdf (and possibly XML export, for Belgium) was added as an attachment of each closing entry generated. This was misleading and error-prone: the attachments only need to be on the the main company's closing.

Moreover, the posting of those entries was fully decoupled: so there was no guarantee all of them will be posted at the same time, leading to the possibility that the tax lock dates are not all set at the same time, and some entries impacting the tax reports are added between the posting of each closing. In that case, right the pdf/export file was the one on the last closing posted, all the other ones were outdated. Again, this could cause confusion, and potentially cause big mistakes.

To handle both those cases, we now enforce the order in which closings need to be posted: the main company needs to be the last one. Posting the main company's closing move hence posts the closing of every sub-companies directly. Then, attachments are only generated on the main closing.

X-original-commit: 1f72fc21fc7d3be026881f2736a78531c7751d4a
Part-of: odoo/enterprise#53176