[REF] account_reports: tax closing: remove useless call to sudo()
Companies linked to the tax unit don't need to be referred to with sudo().

The only case where it could seem useful is when the user normally does not have access to some of these companies, but it actually wouldn't make sense, since the following _get_options would anyway only consider the ones that are actually accessible to the user (which is what we want, else it would create security issues, as it would then be possible to retrieve data from unauthorized companies).

Part-of: odoo/enterprise#52706