[IMP] l10n_nl_reports{_sbr}: round down tax report values
The Dutch government allows to round the amounts to report to them
in the "submitter's advantage" (round down for amounts to pay and up for
amounts to get back). As we have the choice for the way we round, we
decided to only round down.
For more consistency we now include the options (default true) to round
the values in the Dutch tax report UI.

closes odoo/enterprise#53885

X-original-commit: 00ae6efe1e029c157aea3281c72a9728f3589388
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>