[FIX] currency_rate_live: make xe.com default for India again
During the forward-port of [1] to 17.0 [2], I resolved a merge conflict
in the `res_config_settings.py` file incorrectly, which made it so that
xe.com is not the default provider for India anymore (like in previous
versions).

This change fixes that and makes it again the default provider for
India.

[1] 24012f9db268a7722275a508374498bb18bbebdc
[2] 40d62fb3ce62a22e455ec98f2fe692286138786e

closes odoo/enterprise#52917

Signed-off-by: John Laterre (jol) <jol@odoo.com>