[FIX] sale_subscription: fix the visibility for upsell quotations button
As of now, the upsell quotations stat button in the subscription module's sale
order form view is only visible when the sale order is in the draft state.

steps to reproduce:
- upsell a so from the subscription module
- click on the upsell quotation stat button and change the state=sent quotation
- the upsell button will disappear from the confirmed sale order form view

With this commit, the upsell quote stat button will remain visible even when
the sale order is in the sent state.

task-3593973

closes odoo/enterprise#53138

X-original-commit: c92779b44a792813e09d06183b97bc4a5bf9cd5e
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>