[FIX] planning: correct tz test
`test_duplication_should_preserve_local_time` calls
`action_copy_previous_week` wit a timezone in the context.
The issue is that the method get the timezone from the employee or else
the context, yet the employee has a tz 'UTC', so it crashes.

We this commit, we set the desired tz on the employee.

closes odoo/enterprise#53020

X-original-commit: 0781061220ddbdea126fe33fbf53d9372ba5f047
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>