[FIX] appointment: free if event starts yesterday
Current behaviour:
When an employee places an event on a certain day
starting at 00:15 to 18:00 on a brussels TZ,
the event is stored on the UTC TZ, so starting
the day before at 23:15 to 17:00.
This behavior shows the employee as available.

Expected behaviour:
The employee should not be available even
if the event start on another day.

Steps to reproduce:
1. Employee and db has a Brussels TZ
2. Go to Calendar
3. Add an event when employee is available
4. Event starts at 00:15 and ends at 18:00
5. Go to website > Appointment
6. Select Mitchell Admin
7. Look at the date, should be unavailable
8. => Still available

Cause of the issue:
https://github.com/odoo/enterprise/blob/d6135811316837d6523e59163739fc9b954be2e6/appointment/models/calendar_appointment_type.py#L295
If the event starts on another day than the slot,
it isn't taken into account.

Fix:
Adding events from yesterday and tomorrow to know
if they overlap with the slot

opw-3599209

closes odoo/enterprise#57396

X-original-commit: e4137a5f09ce143d0d771827f706ab4cc930ea88
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Co-authored-by: Boulif Nasreddin <bon@odoo.com>