[REF] hr_referral,hr_appraisal: remove _read_group_check_field_access_rights()
Since _read_group_check_field_access_rights() is removed in the
community version, then remove the remaining occurrences of it.

- For hr.referral: Simplify each check by overriding only
check_field_access_rights().
- For hr.appraisal: We actually want security checks on fields depending
on the record itself. Then change the paradigm with 2 new
compute/inverse fields that manage the security during the read and
the write.

closes odoo/enterprise#53409

Related: odoo/odoo#146438
Signed-off-by: Raphael Collet <rco@odoo.com>