[FIX] planning: pop views-specific context keys from dialog context
`highlight_planned` and `highlight_conflicting` are context keys
that are use to highlight certain shifts in gantt and calendar views.
In other views, we use them to hide the other ones.

This context is passed in `getDialogContext`, and it computes a domain
in `load`, which will be an argument of `webSearchRead`.

Yet, obviously, when the user wants to plan a shift,
it makes no sense to present him the shifts in conflict
nor the shifts that he just planned.

So, to avoid having this undesired domain in `webSearchRead`,
we delete both keys from the context we return from `getDialogContext`.

task-3555519

closes odoo/enterprise#49104

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>