[IMP] l10n_be_soda, l10n_be_codabox_bridge: Small improvements of SODA imports.
Problem
---------
With the integration of Codabox in l10n_be_codabox, we need to managed
the automatic import of SODA. This is currently done but can be
improved.

Objective
---------
Adapt the following point:

1. Put the date from <GenDate> by default in the date of document
2. Add the period linked to the document in the ref of the document.
This info is store in the field <AccountPeriod>
3. In the matching form between SODA account and Account in the company:
Merge the account and the name from SODA (to be consistante with the
 column in which the user should add account)

Solution
---------
1. Simply add the date data from the XML in the dictionary that will be
passed to the wizard creating the account move and make sure the wizard
gives that value in the move create method.
2. Concatenate the formatted <AccountPeriod> text to the ref.
3. Add a compute method to computes the display_name for the current
model as it is done in account_account. Use the display name in the
wizard view in readonly instead of the SODA account codes and names.

task-3615994

closes odoo/enterprise#55667

X-original-commit: 4bffd2d0214bed544d12b7320dfbd25bf856efb3
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>