[FIX] pos_iot: Correctly delete linked pos_printer to iot_box
Current behavior:
You couldn't delete an iot_box if a pos_printer was linked to it.
You always get a validation error saying that "Proxy IP cannot be empty"
To fix this we first want to check that the device_id linked to the
printer still exists. To do this we need to move the constraint to the
pos_iot module (because the device_id field is defined there)

Steps to reproduce:
- Install pos_iot
- Activate preparation printer in POS
- Create a printer and link it to an iot_box device
- Try to delete the iot_box

opw-3597370

closes odoo/enterprise#54890

X-original-commit: 950a5b4b143bcd3dbda2f4376f61466768223d4d
Related: odoo/odoo#150537
Signed-off-by: Robin Engels (roen) <roen@odoo.com>