[FIX] marketing_automation: correct allowed_parent_ids in campaign form
Recent changes in how the ORM works in 17.0 have made it so when computing
`allowed_parent_ids`, the <activity.campaign> record of the new activity
is a NewID record instead of a regular record in prior versions.

Additionally this seems to lead to the campaign not being aware of its stored
`marketing_activity_ids`, which means they
cannot be added to `allowed_parent_ids`
which means they are then filtered out by the domain in the form view
of activities.

A temporary solution is to use the property `_origin` to force the ORM
to use the current record instead of what it believes is the new one.

task-3643719

closes odoo/enterprise#53140

X-original-commit: 31dc1d2e9985f739fed52b3c04459ed9068acda5
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>