[FIX] account_reports: totals below section redundancy display
When totals below section option is set, there is a redundancy
in the display of data. Indeed, the amount will be displayed
once on the line itself and once again in the total line.

For now on, if the total below section is set and the total line
is displayed (unfolded), the amount will only on the total line.

task-3642826

closes odoo/enterprise#55416

X-original-commit: 78a01e76cc2c5d0680bab3c02d59159d03587b9f
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>