[FIX] account_reports: tax closing: don't mix companies when checking for advance VAT payments
Before this fix, when doing a tax closing with multiple companies setup as branches, posting one of those closings and then regenerating the others, the tax receivable/payable line of the posted move was considered as an advance tax payment in the other ones, causing their lines to be wrongly computed. This was due to the fact no filtering was done on the company when getting such advance payment, which was plain wrong. This error was shadowed in the past due to the fact accounts were never shared between companies. This is now a bit more complex with branches.

closes odoo/enterprise#53176

X-original-commit: 654025f11ae12f2ab17e223c5aa2f98b9469ddc2
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>