[FIX] l10n_be_reports: Fix EC sales report reminder behavior.
Problem
---------
Since saas-16.4, the banner reminding a user to submit EC Sales Report
was showing when not relevant sales had occured. Furthermore, it would
hide when some of those sales were made.

Objective
---------
Have the opposite behavior.

Solution
---------
Invert the condition that would show or not the banner by remove the
'not'.

task-3645189

closes odoo/enterprise#53235

X-original-commit: 6b93a71b51d94bbb5a2c9d3e3e3e366014e51731
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>