[FIX] l10n_ee_reports: don't display special code on single tax amls
Currently, in the KMD report,  the special code 3 appears on invoices that have a single tax with a single tax amount.
This code should only appear if there are several tax rates on a single account move line. Extra taxes are fetched in the sql query. This commit fixes that issue.

X-original-commit: 368992db03acf52ff1d7870dae5c60b1bedc0b17
Part-of: odoo/enterprise#53321