[FIX] l10n_ee_reports: don't display special code on single tax amls
Currently, in the KMD report,  the special code 3 appears on invoices that have a single tax with a single tax amount.
This code should only appear if there are several tax rates on a single account move line. Extra taxes are fetched in the sql query. This commit fixes that issue.

X-original-commit: 0e2650356196d3d4b0a893b5efda79d71384feb6
Part-of: odoo/enterprise#53343