[FIX] spreadsheet_edition: adjust multi-record selector dropdown styling
Previously, the multi-record selector dropdown occupied an entire line.

This commit addresses this issue by applying the `flex: 1 0 1rem;`
property to the auto-complete dropdown. This adjustment confines the
dropdown to the necessary space. If the space available is less than 1rem,
it will wrap to a new line.

Task ID: 3607067

closes odoo/enterprise#53071

Related: odoo/odoo#146919
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>