[IMP] *: Better invoicing, taxes and company/report config
*: account_intrastat, sale_subscription

This commit serves the purpose of enhancing the invoicing UX and UI.

This is done by various improvements making the invoicing clearer and easier.

Invoices:

The width of the intrastat and the product country column have been set to
fixed values instead of using default ones. This is done to avoid big white
spaces on the view.

Mobile version of Invoices:

The intrastat and subscription fields are only shown in debug mode.

task-3642419

Part-of: odoo/enterprise#61327