[FIX] l10n_de_reports: DateV export several lines into 1
'Buchungstext' is supposed to hold the line label and not the move
name (or ref).
It implies that the idea of grouping the data by accounts + partner
does not hold anymore (as we need to split by line label).
We get rid of this grouping as it's a byproduct of trying to group
the line of product with the line of tax and the receivable/payable.

opw-3633926

closes odoo/enterprise#53757

X-original-commit: aa5b93c6ac9f23ad6c63d5af629da771dc09223e
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>