[IMP] account_reports: add more views to the deferred audit
Currently when auditing entries in the deferred report, you get a list
of AMLs (journal items), but can't change the view. It is useful though
to use other views, like the pivot view e.g.

For this reason we let the user choose to switch between any view
available in the generic Journal Items list.

(Feedback from TSB)

closes odoo/enterprise#60793

X-original-commit: f7ea0e9ea3948aa9896f6457f64a15b3aa033f6f
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>