[FIX] account_consolidation: prevent too-long account name
Issue: Requesting consolidation report for consolidation name that is more than 40 characters resulted in runtime error

Steps to reproduce:
1- Install consolidation app
2- Create a new consolidation
3- In the account mappings of the consolidation, create a mapping to any account with a name longer than 40 characters
4- click the smart button "Consolidated Balance"

Solution:
Issue was that while generating the consolidation report for long name we needed to check if it is to be exported to pdf or xml or just a normal request and no export. This was done by checking the options dictionary but using a key named 'print_mode' this key no longer exists and has been replaced with 'export_mode'.

opw-3629651

closes odoo/enterprise#54365

X-original-commit: d94f48335346680f6935f8f1a02b58d6508929da
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Omar Abosamaha (abom) <abom@odoo.com>