[IMP] l10n_mx_edi: flag "to check" on issues with import
When we're importing a Mexican vendor bill and we're unsure about
something, like a tax that couldn't be found or multiple possible taxes,
a message is logged in the chatter of that invoice in Odoo. However, it
is not easy to identify which imported bills have errors later on when
we have a bunch of them.

In order to make that easier, we flag the vendor bill as "to check"
whenever we encounter an uncertainty at import, so the user can easily
filter on that to follow up on these invoices.

task-3590442

closes odoo/enterprise#53128

Signed-off-by: Laurent Smet (las) <las@odoo.com>