[FIX] hr_expense_extract: don't fill field if there is no value
When the OCR isn't able to detect a value for a field, we shouldn't try
to fill it in.
For most fields, it wasn't much of an issue, but in the case of the
currency, it was very problematic as we're using the `ilike` operator
with an empty string, it matches any string !
Anytime the OCR wasn't able to detect the currency, it could set any
currency in the expense.

closes odoo/enterprise#53431

X-original-commit: 8530be30b145ae6ef55386544a27f0cdf797793d
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>