[REF] *: router service is no longer a service
This commit aims at simplifying the way developers can access
router related information in the webclient. Before this commit,
in components, they had to import the hook `useService` and in
the setup do something like `this.router = useService("router")`. In
services, they had to explicitly declare the dependency to the
router service. In any other functions (like helper functions),
it was a bit painful: the function had to take an env (or the
needed information directly) in arguments, which wasn't ideal.

This commit makes accessing router information easier. The router
service has been removed. Instead, we introduce a `router` object
which can be imported directly from everywhere. Something
similar has already been done for the rpc and orm services.

Part of task 3621046

closes odoo/enterprise#53127

Related: odoo/odoo#146960
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>