[FIX] website_sale_stock_renting : Security time warning
Steps to reproduce:
	1- Install Website, Rentals, Sales and Inventory modules
	2- Set a security time in one of the rentals
	3- Go to Rental settings and set 'Minimum Rental Days' to 1 Day or Delete the Hourly rental option from the product
	4- Go to Website and try to rent the product and set the start date to be before the security time
	5- The 'Your rental product cannot be prepared as fast, please rent later.' will not appear

Current behavior before PR:
When trying to rent a product through Website and the product has a security time if the 'Minimum Rental Days' is set with any duration that is not hours or if there is no Hourly rental option the security time warning will not appear. This happens because the condition that need to be passed to print this warning check if the duration is in hour or not which I think is not necessary because the security time has nothing to do with the UoM of the duration. This condition is only important to check if we will display hours option in the calendar widget or not.

Desired behavior after PR is merged:
This has been solved by removing the condition that checks if the duration's UoM is hours or not. Now we are just comparing the rental start date and the security time.

opw-3553870

closes odoo/enterprise#53372

X-original-commit: 3d427558adf451b03ce1c7a64a5508812f5a8a16
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>