[FIX] l10n_lu_reports: fix custom appendix line name
Line 42 in Appendix A is supposed to allow entering a custom name of an operational expense and its value, both of which are manual values to be entered by the user. The name field is supposed to be a string but currently it's a float as the report doesn't specify that it should be a string.

closes odoo/enterprise#53145

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>