[FIX] timesheet_grid: validate button is missing in pivot view
Before:
In the modified code, parts of the PivotController have been moved to
PivotRenderer. However, the 'button template' is not extended in the pivot view,
leading to issues rendering the validate button.

After:
By extending the correct class (PivotRenderer instead of PivotController) in the
`TimesheetValidationPivotController`, we now have access to methods that were
previously unavailable. Additionally, extending the `buttonTemplate` allows us
to access the validate button, resolving the rendering issue. This modification
ensures that the TimesheetValidationPivotController now extends PivotRenderer,
allowing it to access the necessary methods and resolve the rendering issue for
the validate button.

Task-3623493

closes odoo/enterprise#57591

X-original-commit: 25ce8b166d38d919f29ed24a4a911edd3363d2a6
Related: odoo/odoo#155547
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>