[FIX] l10n_mx_edi: lxml exception when filestore is unavailable
In `_l10n_mx_edi_cfdi_invoice_get_payments_diff()`: If the filestore is not
available (e.g. during upgrades or for any other reason), `ir_attachment.raw`
computation returns an empty byte string. Passing that to
`lxml.etree.fromstring()`, causes `lxml.etree.XMLSyntaxError: Document is
empty, line 1, column 1`. The upgrade of serveral DBs fails because of this
(tbg-978).

Since in that situation it is unclear why the attachment is empty, it is best
to leave the state as is and continue to the next payment.

closes odoo/enterprise#53377

X-original-commit: 1b5dbd50c3910c9b0340b11dd332ec2663a687f7
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>