[FIX] web_cohort: renderer should always be displayed
This commit fixes an issue with the cohort view buttons which are not
accessible when there is no data to display. This is annoying because it
may happen when one changes the measures of the cohort view to one that
won't display any data. The fix here is therefore to remove the data
condition that enables to show the renderer so that even if it is blurred
by the no content helper, buttons can still be interacted with. This
commit also corrects the css selector which disables interactions with
the renderer when sample data is present.

Steps to reproduce:

-go to any cohort view (ex: Helpdesk -> Reporting -> Tickets Analysis)
-filter records with a giberish filter

task-3626278

closes odoo/enterprise#53248

X-original-commit: 3ba4964ca479fc96354cca1a15b05886025285d5
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>