[FIX] social_twitter: fix double tag due to Twitter API response
Bug
===
When answering someone tweet, the Twitter API adds the users at the
beginning of the message (but not always), even if we already added
them manually (so the API response is different from the created tweet).

To fix that issue, we overwrite the API response with the message we
just sent (and so with the message that has been created on Twitter).

Task-3615860

closes odoo/enterprise#53607

X-original-commit: fec4c64c45909ea4c9aabef96fab0378889b12f3
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>