[FIX] knowledge: handle focus events in behaviors
This commit fixes a bug where the user would click outside the editor space in
Knowledge after editing a behavior and this would not be seen as changes inside
the record.

What is happening is that the editor, and the majority of Odoo, uses the event
*blur* to check if any changes has been made to the document. But this event
does not bubble, in opposition to *focusout*, which means that any blur done on
an element that is lower than the editable would not trigger the *onBlur*
method.

What we have done to fix this is simulate a blur event on the editable when we
focus out of a behavior => mimicking an event bubbling to the editable and thus
triggering the onBlur method.

This enables the editable to handle changes done and to save/discard more
efficiently and robustly.

task-3662914

closes odoo/enterprise#54711

X-original-commit: 06056d96058f7484635e4943d2cbc69424e689a2
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>