[FIX] timesheet_grid: infinite recursion in qunit test
Since [1], the "Check that individual and total overtime is properly
displayed" test patches the `setTimeout` function. This sometimes
leads to infinite recursion when the `multi_tab` service has enough
time to initialize: this service make use of `setTimeout` to call the
`heartbeat` method repeatedly. Moreover, patching the `setTimeout`
method is not a good idea since it makes this method synchronous which
totally changes the flow that is tested.

This PR fixes this issue by using the `contains` helper instead: this
method will wait for the element to be inserted in the DOM and perform
the assertion afterwards.

fixes runbot-24889

[1]: https://github.com/odoo/enterprise/pull/46991

closes odoo/enterprise#53326

X-original-commit: 4f0aefe67042c56e7962f88eda9fbb6db321eadb
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>