[IMP] web_grid: add 'show weekends' to grid view
Currently in calendar views it is possible to show/hide the weekends in the
view using the scale selector. In order to gain space, specially in mobile
devices, it is important to have the same feature in grid views. This commit
implements this feature in grid views. Note that the option to show/hide the
weekends is saved in the local storage and it is not shared with the calendar
configuration.

task-3628487

Part-of: odoo/enterprise#52259