[FIX] documents: resolve duplicate shortcut in pdf manager
How to reproduce:
1. Select a file/doc.
2. click on the split tool button
3. opens the PDF split view
4. press control+k (shows command palette)

Before this commit:
List all the commands, but those which are related to
'DocumentsKanbanRenderer' are not working.

Technical Reason:
PDF splitter is a dialog, the context of the Kanban Renderer
stays in the background, 'Select all' and other command logic
related to the 'DocumentsKanbanRenderer', do not work in
PdfManager because of not having the state of it.

After this commit:
Showing commands related to the PdfManager only.

Task-3623653

closes odoo/enterprise#56631

X-original-commit: 2e98772cdb02cba48157b7309a07494bf89b15a4
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>