[FIX] documents_spreadsheet: fix infinite recursion in tests
Since [1], the `toggleCogMenuSpreadsheet` helper patches the
`setTimeout` function. This sometimes leads to infinite recursion when
the `multi_tab` service has enough time to initialize: this service
make use of `setTimeout` to call the `heartbeat` method repeatedly.
Moreover, patching the `setTimeout` method is not a good idea since it
makes this method synchronous which totally changes the flow that is
tested.

This PR fixes this issue by using the `contains` helper instead: this
method will wait for the element to be inserted in the DOM and perform
the assertion afterwards.

fixes runbot-24889

[1]: https://github.com/odoo/enterprise/pull/41851

closes odoo/enterprise#53285

X-original-commit: 87f42ad889ff6ed81f87d0200d149f896e699a23
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>