[FIX] timesheet_grid: display pager in TimesheetTimerListView
Before this commit and it's community counter part [1], the pager
wasn't displayed in the TimesheetTimerListView. This was due to a
wrong interaction between two calls to useSubEnv and useChildSubEnv
(more information available on [1]). The issue has been fixed
directly in the framework, so this commit only adds a test.

opw 3636182

closes odoo/enterprise#53739

X-original-commit: 856e4a3fac5b36a5cf32b4794b7523ef0af2c418
Related: odoo/odoo#148355
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>