[FIX] sign: fix errors introduced by user service removal
PR [1] replaces the user service by a simple object. As a
consequence, it's no longer necessary to call useService and to
store the returned user object on `this` in the setup. PR [1] thus
removed those assignations. However, we forgot to adapt some places
where that user object was used (typically in patches or extensions
relying on the definition of `this.user` in the inherited class).

In the case of this PR, we don't even need to pass the user context
as context of an orm.call, as the orm service automatically adds
the user context itself.

[1] https://github.com/odoo/enterprise/pull/52372

closes odoo/enterprise#53293

Related: odoo/odoo#147265
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>