[FIX] timesheet_grid: keep the timer button when horizontal scroll is made
Before this commit, when the user horizontally scrolls in the grid view
of `My Timesheets` menu, the grid timer button on each row is hidden,
only the row title is kept.

This commit keeps the grid timer button is kept when the user horizontally
scroll in the grid view of `My Timesheets` menu as it is the case for the
row title.

task-3378510

closes odoo/enterprise#54052

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>