[FIX] mrp_workorder: shop floor register production traceback
To reproduce:
- Create BoM for product with one tracked component that is in-stock
- Create MO for BoM, qty 2
- Open shop floor, click on the stock move for the component on the MO
  and change the quantity to 1, save.
- Run the quick action (click 2 units) on register production for the MO

The problem is caused because the MO will become no longer ready,
which should remove it from the view, but the id is still in the cache.

There are other occurrences of the same issue. For example, this also
happens when using the scrap function from the shop floor.

This fix will remedy this problem in a more fundamental way: we now
accept that cases might occur where some record id was cached, but the
corresponding record is not present in the newly fetched recordset
because it no longer conforms to the current filterset.
We decide to in this case allow the corresponding record to just
disappear immediately instead of waiting for the user to manually click
the refresh button or switch workcenter tabs.

Task: 3651433 (nrs 2 & 4)
X-original-commit: c645937f9fc384c82502347f40950e568205be15
Part-of: odoo/enterprise#54774