[FIX] whatsapp: skip user error when message is sent by cron
When the user confirms the order and a WhatsApp message is sent using cron this
error is produced.

Steps to produce:-

- Install `whatsapp_website_sale`
- keep the customer's phone number invalid (E.g. `+1 555-555-5555`)
- from the website, add some products and proceed to checkout
- try to pay for that order
- the error will be produced as an invalid phone number

Problem:-

When a WhatsApp message is sent using `force_send_by_cron=True` then it must
not raise any user error in between regardless of a single or multiple records.
But for obtaining the value `formatted_number_wa`in `_send_whatsapp_template`,
`raise_exception` is `False` only in the case of multiple records and not in the
case of a single record, even if the record is sent using cron.

Solution:-

`raise_exception` must be `False` if `batch_mode` or `force_send_by_cron` is
`True`. While calculating `formatted_number_wa` using `wa_phone_validation` the
`raise_exception` must be false if the message is sent using cron.

Task - 3662747

closes odoo/enterprise#56430

X-original-commit: c321cfe7598b58279858e4c008146da29bf75a9e
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>