[IMP] industry_fsm_sale: do not set the SOL on timesheets in particular
conditions

Before this commit, when an employee timesheeted in an unvalidated
FSM task which has a mapping for that employee and that no SOL
was found in the SO according to that mapping, the SOL of the timesheet
was set to the SOL of the FSM task by default. Which could lead to
confusion for the user.

Now, the SOL of the timesheet is set to False in such conditions.

task-3510920

closes odoo/enterprise#53493

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>