[FIX] whatsapp: prevent error while sending a forwarded message in whatsapp
When the user sends forwarded message to the WhatsApp business account a log
error occurs and the forwarded message will not be sent to WhatsApp business
account

Error:-"ValueError: Expected singleton: mail.message(195, 193, 192)"

This is because at line [1], we receive multiple parent ids in WhatsApp messages
because the value of "context.get()" is none, so at line [1] searches for the
message whose msg_uid is not set (messages either in queue or failed).

This commit fixes the above issue by adding the condition that checks the ID is
present in the context of the message.

https://github.com/odoo/enterprise/blob/272856993b21d9e7edfb573e33a368f80ba0c9c8/whatsapp/models/whatsapp_account.py#L185

sentry-4745952077

closes odoo/enterprise#55147

X-original-commit: 0881c8743477cad30c8395fe370e2298d29c514f
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>