[IMP] project_enterprise, industry_fsm: make fsm group_expand like project
The goal of task is to review the group_expand implemented field service
to be able to do as we did in project but we should only search in the
FSM tasks and not all the tasks as we did in project.

Rather than overriding `_group_expand_user_ids`, we created a hook
in it `_group_expand_user_ids_domain`, which is overridden in
industry_fsm. If `fsm_mode` is in the context, we add the condition
that `is_fsm` is True in the search.

task-3637742

closes odoo/enterprise#53499

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>