[FIX] l10n_account_customer_statements: date fix
The dates in the customer statements header is using the format 'Y'.
This is an issue as the capital Y is using the ISO year-week calendar,
leading to the wrong year being displayed in the report during some weeks.

To reproduce on runbot, simply open the partner ledger on 2024 and print. The end date should be 31/12/2024 but formatted as 31/12/2025

closes odoo/enterprise#53529

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>