[FIX] l10n_nl_reports: XAF export missing ledger
Before this commit:
The generated XAF file while exporting the general ledger is missing the declaration of the accounts with no lines for the given time period which have an opening balance.
The file contains the information regarding the opening balance for an account that is never declared in the file. This leads to an error for some customers when they send the file to the agency.

After this commit:
All the accounts with an opening balance but no line for the time period are declared in the appropriate section of the XAF file.

opw-3567617
opw-3624000

closes odoo/enterprise#59529

X-original-commit: 8c07b73688d8114a9eee6e65887200c87a5e87f1
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Corentin Thaon (thco) <thco@odoo.com>