[FIX] l10n_be_hr_contract_salary: recompute init value
When an offer is generated from an existing contract, the initial value
for the hospital and ambulatory cost is takend from the previous
contract. If the cost for an adult and/or a children is changed in the
system parameters is changed, it is not taken into account for offers
comming from an existing contract having already insurance amount.

So we recompute the cost of the insurance based on the system config
parameter values.

closes odoo/enterprise#53583

X-original-commit: 4a138a1a607e5c0a182d9b7a1cbd7de5e1ef01e2
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>