[IMP] knowledge: redirect to the parent article when trashing an article
When the user sends an article to the trash, the user will be redirected
to the first accessible article. This redirection can be frustrating when
the user reads article and opens an article item to delete it. In this
scenario, the user will lose the page they were on. To avoid that, the
users will now be redirected to the parent article when archiving an
article. This will allow the users to continue reading the parent article.

In addition to that, this commit will simplify the model `knowledge.article`
by removing redundant actions used to archive/unarchive an article
(i.e: the `action_archive_article` and `action_unarchive_article`).
Those actions will be replaced by the standard actions of the ORM
(i.e: `action_archive` and `action_unarchive`).

It turns out that the actions returned by those API methods was almost
always ignored by the caller. To reduce the number of queries executed
when calling the API methods, those methods will now no longer return
an action. It will be up to the caller to decide what to do after
archiving an article which gives more flexibility and control.

task-3619503

Part-of: odoo/enterprise#51694