[FIX] knowledge: fix undefined userService in main list controller
The user service has recently been removed, but it is still utilized in
the main list controller of Knowledge to verify if the logged-in user
is an admin before displaying the 'duplicate' option. As the service no
longer exists, the service can not be loaded and any function call on that
service will trigger an error. This commit addresses the issue by using
the new `user` object instead of the outdated service.

Steps to reproduce the error:
1. Go on Knowledge
2. Click on the "Search" menu item of the navbar
3. In the list view, select a record by checking a checkbox

=> TypeError: `this.userService` is `undefined`

TO BE: There should be no error when checking a checkbox.

Reference: https://github.com/odoo/enterprise/pull/52372/

task-3663517

closes odoo/enterprise#53552

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>