[FIX] sale_subscription: too much pending payment messages
Before this commit, when the customer payed a subscription and the payment was still pending, three alerts were being showed in the screen alerting that its payment was pending, which was too much.

After this commit, only the payment module pending message was kept: "Your payment has ben successfully processed but is waiting for approval". This way, the information is presented in a cleaner and direct way.

closes odoo/enterprise#54249

Task-id: 3649938
X-original-commit: 4d838446766433f7082e6df22980a5c9ad90ebc3
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>