[FIX] knowledge: remove noisy elements from the export
This commit fixes a bug where the export button would take into account
elements that are useless to print (such as discuss tabs).

Now we added a scss rule inside the `knowledge_print.scss` file to remove any
element that is not inside of the action_manager => all of the discuss
containers, the editor toolbar, etc.

task-3653295

closes odoo/enterprise#53675

X-original-commit: 2c5b74f8cd27fed1eeeeb0ba412b3bd9c26a9b0d
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>