[FIX] pos_blackbox_be: certification update
In anticipation of Version 17 certification, multiple adjustments were
required. The option to reprint tickets has been eliminated from the
tip screen. Furthermore, the functionality on this screen has been
rectified to enable the invoicing of previously non-invoiced orders.
Previously, this action was not feasible when an order was placed using
a blackbox.

This commit also involves alterations to values within the sale details
template. Moreover, steps have been taken to guarantee the precision of
sequences for both pro forma and standard sales.

Changes in functionalities have been made that were previously
restricted:
    - Enabled the option to cancel orders
    - Change quantities of orderlines
    - Add global discout button
    - Remove the work button. Work IN/OUT is now done automatically
    - Add the possibility to cash in/out

Sequence numbers are now clearly indicated in the logs and on the order
receipts.

Community PR: odoo/odoo#148155

Part-of: odoo/enterprise#53654
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>