[IMP] stock_barcode: barcode_allow_extra_product
Add a new field on the picking type, `barcode_allow_extra_product`, to
be able to add extra product when processing an operation of this type
in the Barcode App.
By default, this setting will be set on True, and thus the behavior will
not change comparing to before this commit.
But when this setting is disabled, it won't be possible to scan an
unreserved product for the given operation (the button to manually add a
product will also be hidden.)

task-3650361

Part-of: odoo/enterprise#53610