[IMP] stock_barcode_picking_batch: group lines
Add a new field on picking types, `group_lines_by_product`.
When active, for the batches of this type, the move line will be grouped
by product (at least if they have the same source/destination) like is
already the case for tracked products.

The purpose of this change is to fluidify the batch picking when the
pickers process a batch of pickings adressed to the same partner
(because in such case, it's not usefull to process lines for each
different pickings individually.)

There is a side effect: we no more display a color guide on barcode
lines when this setting is active (because a grouped line can include
lines from different pickings.)

task-3650361

closes odoo/enterprise#53610

Signed-off-by: William Henrotin (whe) <whe@odoo.com>